Add dependencies to concoct_linux.yaml to accomodate concoct v. 1.1
Concoct is a mess. It needs libgsl.so.25 for a C dependency, which isn't included in Conda's Concoct installs. Best workaround so far is to add gsl<2.6 and scikit-learn<1.2 to concoct_linux.yaml. Gsl includes libgsl.so.25 unless it's >= 2.6, when I think it switches to libgsl.so.27. And scikit-learn > 1.1 throws an error because something Concoct uses is deprecated. Tried downgrading concoct to 1.0, but this causes weird bugs in other parts of the pipeline.